REF: collect arithmetic asvs
